+++ gazon-microservice-starter/src/main/java/com/gitlab/exception/handler/GeneralExceptionHandler.java
@@ -28,4 +28,10 @@ public class GeneralExceptionHandler {
     public ErrorResponseDto handleConstraintViolationException(ConstraintViolationException ex) {
         return new ErrorResponseDto(HttpStatus.NOT_FOUND.value(), ex.getMessage());
     }
+
+    @ResponseStatus(HttpStatus.FORBIDDEN)
+    @ExceptionHandler(value = ForbiddenException.class)
+    public ErrorResponseDto handleForbiddenException(ForbiddenException ex) {
+        return new ErrorResponseDto(HttpStatus.FORBIDDEN.value(), ex.getMessage());
+    }
 }