+++ gazon-service/src/main/java/com/gitlab/controller/BankCardRestController.java
@@ -2,7 +2,11 @@ package com.gitlab.controller;
 
 import com.gitlab.controllers.api.rest.BankCardRestApi;
 import com.gitlab.dto.BankCardDto;
+import com.gitlab.mapper.BankCardMapper;
+import com.gitlab.model.BankCard;
+import com.gitlab.model.User;
 import com.gitlab.service.BankCardService;
+import com.gitlab.service.UserService;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -12,17 +16,22 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
+
+import static com.gitlab.util.UserUtils.isAdmin;
 
 @Validated
 @RestController
 @RequiredArgsConstructor
-@PreAuthorize("hasRole('ADMIN')")
 @SecurityRequirement(name = "bearerAuth")
 public class BankCardRestController implements BankCardRestApi {
 
     private final BankCardService bankCardService;
+    private final UserService userService;
+    private final BankCardMapper bankCardMapper;
 
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<List<BankCardDto>> getPage(Integer page, Integer size) {
         var bankCardPage = bankCardService.getPageDto(page, size);
         if (bankCardPage == null || bankCardPage.getContent().isEmpty()) {
@@ -32,36 +41,58 @@ public class BankCardRestController implements BankCardRestApi {
     }
 
     @Override
-    public ResponseEntity<BankCardDto> get(Long id) {
-        Optional<BankCardDto> optionalBankCardDto = bankCardService.findByIdDto(id);
+    public ResponseEntity<BankCardDto> get(Long cardId) {
+        User user = userService.getAuthenticatedUser();
 
-        return optionalBankCardDto
-                .map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
+        if (isAdmin(user)) {
+            return bankCardService.findByIdDto(cardId)
+                    .map(ResponseEntity::ok)
+                    .orElse(ResponseEntity.notFound().build());
+        }
+
+        if (isAuthorized(user, cardId)) {
+            return bankCardService.findByIdDto(cardId)
+                    .map(ResponseEntity::ok)
+                    .orElse(ResponseEntity.notFound().build());
+        } else return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @Override
     public ResponseEntity<BankCardDto> create(BankCardDto bankCardDto) {
-        BankCardDto savedBankCardDto = bankCardService.saveDto(bankCardDto);
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(savedBankCardDto);
+        return ResponseEntity
+                .status(HttpStatus.CREATED)
+                .body(bankCardService.saveDto(bankCardDto));
     }
 
     @Override
-    public ResponseEntity<BankCardDto> update(Long id, BankCardDto bankCardDto) {
-        Optional<BankCardDto> updatedBankCardDto = bankCardService.updateDto(id, bankCardDto);
+    public ResponseEntity<BankCardDto> update(Long cardId, BankCardDto bankCardDto) {
+        User user = userService.getAuthenticatedUser();
 
-        return updatedBankCardDto
-                .map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
+        if (isAdmin(user) || isAuthorized(user, cardId)) {
+            Optional<BankCardDto> optionalBankCardDto = bankCardService.updateDto(cardId, bankCardDto);
+            return optionalBankCardDto.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        } else return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @Override
-    public ResponseEntity<Void> delete(Long id) {
-        if (bankCardService.delete(id)) {
-            return ResponseEntity.ok().build();
-        } else {
-            return ResponseEntity.notFound().build();
+    public ResponseEntity<Void> delete(Long cardId) {
+        User user = userService.getAuthenticatedUser();
+
+        if (isAdmin(user)) {
+            return (bankCardService.deleteDto(cardId).isPresent()) ?
+                    ResponseEntity.ok().build() : ResponseEntity.notFound().build();
         }
+
+        if (isAuthorized(user, cardId)) {
+            return (bankCardService.deleteDto(cardId).isPresent()) ?
+                    ResponseEntity.ok().build() : ResponseEntity.notFound().build();
+        } else return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+    }
+
+    private boolean isAuthorized(User user, Long requestedCardId) {
+        return user.getBankCardsSet()
+                .stream()
+                .map(BankCard::getId)
+                .anyMatch(cardId -> Objects.equals(cardId, requestedCardId));
     }
 }