+++ gazon-service/src/main/java/com/gitlab/controller/PersonalAddressRestController.java
@@ -2,7 +2,8 @@ package com.gitlab.controller;
 
 import com.gitlab.controllers.api.rest.PersonalAddressRestApi;
 import com.gitlab.dto.PersonalAddressDto;
-import com.gitlab.dto.UserDto;
+import com.gitlab.dto.ShippingAddressDto;
+import com.gitlab.model.PersonalAddress;
 import com.gitlab.model.ShippingAddress;
 import com.gitlab.model.User;
 import com.gitlab.service.PersonalAddressService;
@@ -16,10 +17,11 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+
+import static com.gitlab.util.UserUtils.isAdmin;
 
 @Slf4j
 @Validated
@@ -42,19 +44,19 @@ public class PersonalAddressRestController implements PersonalAddressRestApi {
     }
 
     @Override
-    public ResponseEntity<PersonalAddressDto> get(Long id) {
+    public ResponseEntity<PersonalAddressDto> get(Long personalAddressId) {
         User user = userService.getAuthenticatedUser();
-        if (user.getRolesSet().toString().contains("ADMIN")){
-            return personalAddressService.findByIdDto(id)
+
+        if (isAdmin(user)){
+            return personalAddressService.findByIdDto(personalAddressId)
                     .map(ResponseEntity::ok)
-                    .orElseGet(() -> ResponseEntity.notFound().build());
-        } else if (getShippingAddressId(user, id).equals(id)) {
-            return personalAddressService.findByIdDto(id)
-                    .map(ResponseEntity::ok)
-                    .orElseGet(() -> ResponseEntity.notFound().build());
-        } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+                    .orElse(ResponseEntity.notFound().build());
         }
+        if (isAuthorized(user, personalAddressId)) {
+            return personalAddressService.findByIdDto(personalAddressId)
+                    .map(ResponseEntity::ok)
+                    .orElse(ResponseEntity.notFound().build());
+        } else return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @Override
@@ -64,39 +66,36 @@ public class PersonalAddressRestController implements PersonalAddressRestApi {
     }
 
     @Override
-    public ResponseEntity<PersonalAddressDto> update(Long id, PersonalAddressDto personalAddressDto) {
+    public ResponseEntity<PersonalAddressDto> update(Long personalAddressId, PersonalAddressDto personalAddressDto) {
         User user = userService.getAuthenticatedUser();
-        if (user.getRolesSet().toString().contains("ADMIN")){
-            return ResponseEntity.ok(personalAddressService.update(id, personalAddressDto));
-        } else if (getShippingAddressId(user, id).equals(id)) {
-            return ResponseEntity.ok(personalAddressService.update(user.getId(), personalAddressDto));
-        } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
+
+        if (isAdmin(user) || (isAuthorized(user, personalAddressId))) {
+            Optional<PersonalAddressDto> updatedAddressDto = Optional
+                    .ofNullable(personalAddressService.update(personalAddressId, personalAddressDto));
+            return updatedAddressDto.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        } else return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @Override
     public ResponseEntity<Void> delete(Long id) {
         User user = userService.getAuthenticatedUser();
-        if (user.getRolesSet().toString().contains("ADMIN")) {
-            return (personalAddressService.deleteDto(id).isPresent())
-                    ? ResponseEntity.ok().build() : ResponseEntity.notFound().build();
-        } else if (getShippingAddressId(user, id).equals(id)) {
-            return (personalAddressService.deleteDto(id).isPresent())
-                    ? ResponseEntity.ok().build() : ResponseEntity.notFound().build();
-        } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 
+        if (isAdmin(user)) {
+            return (personalAddressService.deleteDto(id).isPresent())
+                    ? ResponseEntity.ok().build() : ResponseEntity.notFound().build();
         }
+
+        if (isAuthorized(user, id)) {
+            return (personalAddressService.deleteDto(id).isPresent())
+                    ? ResponseEntity.ok().build() : ResponseEntity.notFound().build();
+        } else return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
-    private Long getShippingAddressId (User user, Long id) {
+    private boolean isAuthorized(User user, Long requestedPersonalAddressId) {
         return user.getShippingAddressSet()
-            .stream()
-            .map(ShippingAddress::getId)
-            .filter(saId -> Objects.equals(saId, id))
-            .findAny()
-            .orElseThrow(() -> new EntityNotFoundException("The address does not belong to the user.(Адрес не принадлежит пользователю)."));
+                .stream()
+                .map(ShippingAddress::getId)
+                .anyMatch(cardId -> Objects.equals(cardId, requestedPersonalAddressId));
     }
 
 }