+++ gazon-service/src/main/resources/db/changelog/v-1.38-fill_user_id_to_shipping_address_table/01-fill_user_id_to_shipping_address_table.sql
@@ -1,6 +1,0 @@
-insert into users_shipping_address (user_id, user_shipping_address_id) values (1,1);
-insert into users_shipping_address (user_id, user_shipping_address_id) values (1,2);
-insert into users_shipping_address (user_id, user_shipping_address_id) values (2,2);
-insert into users_shipping_address (user_id, user_shipping_address_id) values (2,3);
-insert into users_shipping_address (user_id, user_shipping_address_id) values (3,3);
-insert into users_shipping_address (user_id, user_shipping_address_id) values (4,1);