+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [Unreleased] 
 - #task_number - short task description
 - #176 - Add security in PersonalAddressRestApi
 - #173 - Add pagination to ProductSearchController
@@ -13,6 +13,8 @@
 - #193 - Checking and add uniqueness_restrictions in the database
 - #198 - Add sorting of payments by ID
 - #175 - Updated security
+- #204 - Add a Central Bank client
+- #183 - Add Security BankCard
 
 ## [1.4.0] - 2.04.2024
 - #167 - Testing the conversion of Product review to productDTO rating